Match the corpus to what the model can learn, and stop paging on int64
Three corrections from the v79/v80 post-mortem.

1. Repetition teaches an algorithm; uniqueness does not.

v74's multiplication task scored 0.93 with 712 distinct operand pairs
repeated 56x each over 40,000 rows. The omni force task -- the same
operation -- scored 0.03 with 24,000 combinations appearing ~1.7x each.

V79_OMNI_FRONTIER.md claimed "duplicated rows are precisely what a
recitation-proof benchmark exists to punish". That is wrong for an
algorithmic task: the benchmark draws unseen operands from the same space,
so the model cannot pass by recall and repetition is how it learns the
method. Uniqueness is right for a knowledge corpus, where a repeated fact is
memorised; it is wrong for a procedure. The exhaustion machinery added in
v79 was solving a problem that did not exist while destroying the repetition
that did the teaching. It survives as --unique; the default now repeats and
reports distinct_prompts and repetition so the factor is visible.

2. The decomposition now matches v74's format byte for byte.

v80 added place-value working to the multiplications. Its step-6,000 probe
was unchanged, but the mechanism moved: the partial products became correct
and the failure moved to the running sums, which I had written as
single-step four-digit additions -- the same mistake one level up.

The fix is not to decompose the sums; that costs +23 tokens, pushes the max
past the 128-token sequence length and forces it to 192 (~50% more training
time). Matching v74 costs nothing:

    80 x 3  ->  80 x 3 = 240, 0 x 3 = 0
    25 x 7  ->  20 x 7 = 140, 5 x 7 = 35

Two details corrected: v74 keeps the zero place, so every two-digit problem
has exactly two partial products and one shape; and v74 writes no addition
step, going straight to the total. Operand ranges narrow to v74's proven
11-99 x 2-9, which narrows the task definition and the benchmark with it,
exactly as v74's own multiplication task is defined and scored.

Measured mid-run: v80's step-9,000 probe reads 0.15 against v79's 0.09, so
the decomposition is working directionally even before these corrections.

3. The corpus was stored at four times the size it needed.

Flagged in V75_CRASH_RECOVERY.md and deferred as speculative. It stopped
being speculative when v79 spent hours at 17.12 s/step against a 4 s/step
norm -- 25.6 GB committed on a 15.6 GB machine, the trainer holding 4.44 GB
but resident in 1.2 GB, faulting its own corpus back from the pagefile.

Token ids below 9,000 were int64: 1.78 GB of that 4.44 GB. compact_dtype
picks the narrowest type holding the vocabulary -- int16 for everything here,
int32 as fallback, never a silent overflow -- with the limit at 32,000 so the
-100 ignore label stays representable. 1.78 GB -> 0.44 GB. Batches widen to
int64 on the way into the model, 16 x 128 values per step.

A 20-step end-to-end smoke run produced byte-identical losses (3.9684,
3.4407, 3.0595, 2.8853) and then crashed in the report phase:

    RuntimeError: ... but got torch.ShortTensor instead
    (while checking arguments for embedding)

routing_report was a missed consumer. Unit tests would not have caught it --
it only runs at the end of a full run -- and the crash supervisor could have
restarted v80 straight into it.

Tests: 2,961 passed, 2 skipped.

+++ source/train_mimomix_generalisation.py
@@ -615,8 +615,15 @@ def run(args: argparse.Namespace) -> Dict[str, Any]:
     for step in range(args.start_step + 1, args.steps + 1):
         model.train()
         pick = torch.randint(0, train_x.shape[0], (args.batch_size,), generator=generator)
-        batch_x = train_x[pick] if accelerated and args.resident_corpus else train_x[pick].to(device)
-        batch_y = train_y[pick] if accelerated and args.resident_corpus else train_y[pick].to(device)
+        # The packed corpus is stored in the narrowest integer type that holds
+        # the vocabulary (see `mimomix_text.compact_dtype`), which is what
+        # keeps a 900k-row corpus off the pagefile. Embedding lookup and
+        # cross_entropy both need int64, so the batch -- 16 x 128 values -- is
+        # widened here rather than the whole corpus being held wide.
+        batch_x = (train_x[pick] if accelerated and args.resident_corpus
+                   else train_x[pick].to(device)).long()
+        batch_y = (train_y[pick] if accelerated and args.resident_corpus
+                   else train_y[pick].to(device)).long()
 
         if autocast_dtype is not None:
             with torch.autocast(device_type=device.type, dtype=autocast_dtype):
